Renamed module to createRouter
